Change subscription domain object prices to floats
The conversion to an int with precision of 2 decimals is now handled inside the get/set attribute methods of the subscription model

+++ app/Domain/Subscriptions/Subscription.php
@@ -13,9 +13,9 @@ final class Subscription
         public readonly string $name,
         public readonly string $description,
         public readonly Currency $currency,
-        public readonly int $price,
+        public readonly float $price,
         public readonly BillingInterval $billingInterval,
-        public readonly ?int $fee,
+        public readonly ?float $fee,
     ) {
     }
 }
